Migrate bot to Aiogram 3.x structure
