+++ .env.example
@@ -1,0 +1,1 @@
+BOT_TOKEN=YOUR_BOT_TOKEN_HERE
