--- bot/token_manager.py
@@ -1,28 +1,0 @@
-import secrets
-from datetime import timedelta
-
-# Durations en días
-DURATIONS = {
-    '1d': 1,
-    '1w': 7,
-    '2w': 14,
-    '1m': 30,
-    'forever': 3650  # Aproximadamente 10 años
-}
-
-
-def generate_token(duration_key: str) -> tuple[str, int]:
-    """Genera un token y devuelve el par (token, dias_de_duracion)."""
-    duration = DURATIONS.get(duration_key, 1)
-    token = secrets.token_urlsafe(8)
-    return token, duration
-
-
-def create_token(duration_key: str):
-    """Create and store a token, returning it and its duration."""
-    from .database import save_token
-
-    token, duration = generate_token(duration_key)
-    save_token(token, duration)
-    return token, duration
-
