--- bot/utils/config.py
@@ -1,39 +1,0 @@
-import sqlite3
-from typing import Optional
-
-DB_NAME = "subscriptions.db"
-
-
-def init_config():
-    conn = sqlite3.connect(DB_NAME)
-    c = conn.cursor()
-    c.execute(
-        """
-        CREATE TABLE IF NOT EXISTS config (
-            key TEXT PRIMARY KEY,
-            value TEXT
-        )
-        """
-    )
-    conn.commit()
-    conn.close()
-
-
-def save_config(key: str, value: str) -> None:
-    conn = sqlite3.connect(DB_NAME)
-    c = conn.cursor()
-    c.execute(
-        "INSERT OR REPLACE INTO config (key, value) VALUES (?, ?)",
-        (key, value),
-    )
-    conn.commit()
-    conn.close()
-
-
-def load_config(key: str, default: Optional[str] = None) -> Optional[str]:
-    conn = sqlite3.connect(DB_NAME)
-    c = conn.cursor()
-    c.execute("SELECT value FROM config WHERE key=?", (key,))
-    row = c.fetchone()
-    conn.close()
-    return row[0] if row else default
