+++ handlers/start.py
@@ -1,0 +1,9 @@
+from aiogram import Router
+from aiogram.filters import CommandStart
+from aiogram.types import Message
+
+router = Router()
+
+@router.message(CommandStart())
+async def cmd_start(message: Message):
+    await message.answer("Bot is running!")
