+++ requirements.txt
@@ -1,2 +1,2 @@
 aiogram>=3.0.0
-
+python-dotenv>=1.0.0
